requirementx.txt
asgiref==3.8.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
mysqlclient==2.2.4
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycryptodomex==3.21.0
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzipper==0.3.6
requests==2.32.3
SQLAlchemy==2.0.37
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
